Delete the wire-package copies that lagged the server contract
#84 added tckdb_schemas/workflows/network_pdep_upload.py and
tckdb_schemas/transport.py as near-copies of the backend schemas rather than
shims, and they immediately fell behind: the PDep copy carried none of #82's
blocking n_imag checks (13 stationary_point references in the backend schema,
0 in the copy) and the transport copy lacked #78's
validate_has_scientific_content.

Nothing outside their own test imported them and neither was exported from the
package root, so no behaviour changes here. The hazard was prospective and
worse than a missing module: they ship in the package a client installs, so a
client validating against them would be told a payload is fine that the server
then refuses -- a confident wrong answer instead of an import error.

Deleting rather than shimming because nothing in the tree stated an intent to
migrate: no shim-identity test, no deprecation note, no importer. A copy is
only safe when something makes drift impossible, and that machinery was never
written.

Two tests in the deleted test module had no backend equivalent and are ported
rather than lost:

  - test_upload_schema_exposes_no_fk_ids_or_hashes guards a project-wide rule
    (uploads carry local keys and scientific content, never our primary keys),
    not a PDep detail.
  - test_chebyshev_grid_dimensions_must_match_declared_orders: a coefficient
    grid contradicting its declared shape cannot be evaluated at any (T, P).

Its other unique cases cover the energy_transfer coverage and scope rules,
which are being redesigned; they are re-established there rather than ported
into a shape about to change.

backend PDep suite 65 passed; schemas package 38 passed.
